requirements.txt
Flask
feedparser
google-generativeai
python-dotenv